Add two ifc files and generate links in the Getting Started page
Feedback: Add two example files to the documentation on GitHub: one that only contains a static system and one that has the results added. If someone comes across IfcTruss and wants to get a first impression, they can use the files directly without having to use the Python code first.